updated modules, pictures, logic
