+++ README.txt
@@ -1,0 +1,35 @@
+The Oracle
+
+
+I wanted to create magic 8 ball themed page which would answer to different questions.
+The idea for this came perhaps from Codecademy's lessons but also I remember one site
+called Lintukoto from my childhood. It had simple style but it was possible to ask
+questions and you could see all of them and the answers on some container. The page 
+had also some other simple games. 
+
+I have been building this page and improving it, adding better logic to it, make 
+it anwser to different question types and it also has two small text adventures.
+
+Also, I found out how to add custom domain to GitHub pages, how to add https instead of just http 
+and some other new things to me. 
+
+Now(4.1.2024) I am coming back to see my code and I am seeing lots of things that
+could have been done easier way, but that is the fun of working with a project and improving and learning! 
+The question is, do people normal modify the code or just start over with the parts that
+are working smoothly..
+
+I did use Youtube, some pages and GitHub Copilot and ChatGPT making this site at the beginning, but
+nowadays I like to achieve things more on my own, but these 2 tools provide sometimes valuable information.
+Also I got help with how to make the eyes move and put modified sprite over it.
+
+
+
+Things to add:
+  
+-Make the code more readable
+-Put picture hint of the text adventures
+-Music
+
+----------------
+Oracle and dragon sprites: made with free tool.
+Other sprites by me.
