+++ content.mjs
@@ -1,0 +1,265 @@
+const answers = {
+    why:  [
+        'Because.',
+        'Because of reasons.',
+        'Because of my cat. He has a cool name.',   //CAT
+        'Because I am the Oracle.',
+        'Because of one hairy gnome.',          //5
+        'Whatever reason.',
+        'Why you, why me, why anything?',
+        'Because you are curious.',
+        'Because you are excited.',
+        'In the dance of possibilities: Why not?',                 //10
+        'Because of the weather.',
+        'Because the ancient scrolls foretell it so.',
+        'You sure alredy know why.',
+        'Because it is just awesome.',
+        'It is not always so simple.',           //15
+
+
+
+],
+    how:  { 
+        general: [
+        'How? No idea, but do you know the name of my cat?',           //CAT
+        'Hmm, how... do not ask me. Ask my cat.', //CAT
+        'Like the wizard looking guy next door.',
+        'Like Gandalf would.',
+        'Like a boss dragon.',             //5
+        'Like the witch looking woman next door.',
+        'With a great attitude.',
+        'With an akward smile on your face.',
+        'I have an idea, but I cannot tell you.',
+        'With a little help from your friends.',           //10
+        'Try the direct approach and the indirect approach after that.',
+        'With the magic of positivity.',
+        'Too shy to tell you.',
+        'With strong arms and a brave heart.',
+        'I might know some magic but how would I ever know that?',
+        'Just push it through.', //15'
+    ],
+
+
+
+    "how many": [
+        'How many? Too hard to count.',
+        'How many? Multiple it by 2 and add 3.',
+        'How many? Not enough.',
+        'How many? Maybe something between 1-150.',
+        'How many? Not so many.', //5
+    ],
+
+    "how much": [
+        'How much? I do not have time to count.',
+        'How much? What if it is too much?',
+        'How much? Not enough.',
+        'How much? A lot.',
+        'How much? Not so much.', //5
+    ], 
+    "how are you": [
+        'I am fine, thank you. But my cat is missing.', //CAT
+        'I am actually doing great, but my cat is missing.', //CAT
+        'Well, everyday is a blessing.',
+        'I am doing great, thank you. How are you?',
+        'Healthy as a horse, thank you. How are you?', //5
+
+    ], 
+    "how old": [
+        'Age is just a number.',
+        'Old enough to know better.',
+        'Young enough to have fun but too naive to know better.',
+        'Old enough to be wise.',
+        'Young enough to be curious.', //5
+
+    ],
+    "how often": [
+        "How often? Not so often.",
+        "How often? I should not tell you.",
+        "How often? I am not the right person to ask.",
+        "How often? What do you think?",
+        "How often? Ask your mother.", //5
+    ],
+
+    "how about": [
+        "How about 'no'?",
+        "How about we try to ask better questions?",
+        "How about we try to find my cat?", //CAT
+        "How about we try to find the meaning of life?",
+        "How about we try to find the best pizza in the world?", //5
+    ],
+    "how dare you": [
+        "How dare you?",
+        "I let this one speak for itself.",
+        "I dare you to ask a better question.",
+        "..." ,
+        "Stop whining.", //5
+    ],
+    "how long": [
+        "Long enough.",
+        "Too long.",
+        "No idea. But how long to the next meal? I am hungry.",
+        "How long? Is that a name?",
+        "Something between 2 hours, 200 km and a one potato." //5
+    ],
+
+},
+    "who": [
+    "Your mama.",
+    "The person you least expect.",
+    "A hairy gnome.",
+    "Gandalf the White.",
+    "A dragon.", //5
+    ],
+    what: [
+        'You know what? My cat might know.', //CAT
+        'No idea. But do you know the name of my cat?', 
+        'What?',
+        'It is a mystery...',
+        'You already know the answer.', //5
+        'What?',
+        'Explore and you will find out.',
+        'The answer is in your heart.',
+        'I will let this one speak for itself.',
+        'Could be anything.', //10
+        'Could be 42',
+        'Do I look like I know? Ok, I know. Hehe.',
+        'This and that.',
+        'That is for you to discover.',
+        'There is not easy answer to "what".', //15
+
+
+
+    ],
+    when: [
+        'Just when you know the name of my cat.', //CAT
+        'When my cat is back here.', //CAT
+        'When you hear my cat purring.',
+        'When the time is right.',
+        'When the night falls.', //5
+        'When you are least expecting it.',
+        'When you are giving your best.',
+        'when you are not giving up.',
+        'When you are not afraid.',
+        'When you are being confident.', //10
+        'When you hear a hairy gnome singing by the river.',
+        'When the unicorns meet.', //UNICORN
+        'After you have finished your vegetables.',
+        'When you are being positive.',
+        'When you are being optimistic.', //15
+    ],
+    where: [
+        'No idea. But do you know the name of my cat?', //CAT
+        'Where the light touches the ground.',
+        'No idea, but where this is: "second star to the right, and straight on till morning.',
+        'Hmm... under the rainbow?',
+        'Your dog sure knows the answer.', //5
+        'Just try to remember.',
+        'Under the sea. No, just follow the fairy lights.',  //musicnotes?
+        "Search anywhere, just don't go to the dragon's lair.",
+        'In the land of possibilities.',
+        'Right here.', //10
+        'Sorry, I do not have a map.',
+        'In the land of the fairies.',
+        'In the next village.',
+        'In the next festival, ask the fortune teller.',
+        'Here and there, who knows?', //15
+
+
+
+],
+    "am I": [  //ottaakin tavun mukaan, esim. "name" johti tänne, siksi "am I"
+        'My cat with a great name likes you, so you must be ok.', //15 //CAT
+        'Nobody is perfect, except my cat with a great name.', //CAT
+        'We think you are doing great.',
+        'You are awesome, like a fancy glitter unicorn.',  //UNICORN
+        'Get yourself together and you will be fine.', //5
+        'Yes, in a positive way.',
+        'Am I?',
+        'I would change my perspective to more positive if I were you.',
+        'Yes, like a phoenix.',
+        'Good job, you are getting better every day.', //10 
+        'Things are getting better.',
+        'Too personal question, I do not want to think about it.',
+        'If you mean the opposite of negative, then yes.',
+        'If you mean the opposite of positive, then no.',
+        'Yes you are, in a positive way.', //15
+
+    ],
+    if: [
+        'In that case, do not worry. My cat does not either.', //CAT   
+        'Then you can pet my cat.', //CAT
+        'It is ok, it will go well.',
+        'In any case, you have to trust your intuition.',
+        'If this happens, you will figure it out.', //5
+        'Do not worry, you just have to trust yourself.',
+        'Well, then it is not meant to be.',
+        'In this case, bring your umbrella.',
+        'You have to trust the journey.',
+        'That will not happen.', //10
+        'Have you ever heard the sentence:"Hakuna Matata"?',
+        'Then double check your sources.',
+        'In that case, remember to buy some cookies.',
+        'If that would be the case, I would be very surprised.',
+        'In that case, KEEP CALM AND TROLL ON.', //15
+    ],
+    should: [
+        'No. But you should know the name of my cat.', //CAT
+        'Perhaps you should.',
+        'Maybe you should, maybe you should not.',
+        'Is it a recommendation or is it optional?',
+        "'Should' is a strong word.",               //5
+        "There are not many 'shoulds' in the world.",
+        "Should you? Should I?",
+        "Maybe. I am not sure about the outcome, but it will be fun!",
+        "If it's going to improve your life, then you should.",
+        'If it is going to help someone, then you should.', //10
+        'But does it help anyone?',
+        'So many shoulds, so little time.',
+        'Only if it is a morally right thing to do.',
+        'You know, backwards "should" is "dluohs".',
+        'What do you think? Should you?', //15
+
+    ],
+
+    "do you": [
+        'Not me, but my cat with a great name is a good listener.', //CAT
+        'I do not. But do you know the name of my cat?', //CAT 
+        'Thanks for asking, but I am not sure.',
+        'Me? I am just a simple Oracle.',
+        'Yes, yes, very funny. Next question.', //5
+        'Haha, you are so funny.',
+        'Let me ask: do you?',
+        'Is there anybody who does not?',
+        'Occasionally.',
+        'Who? Me? Pffft.',  //10
+        'I know all the answers ecxept this one.',
+        'Are you really asking me?',
+        'Better not tell my personal secrets.',
+        'Do I what? I do not understand.',
+        'I would not advise anyone to admit that.', //15
+
+
+    ],
+    general: [
+    'No idea, but do you know the name of my cat?', //CAT
+    'The answer is... umm, how should I read this?',
+    'Not so sure, but my cat might know.', //CAT
+    'Well... I would think about it.',
+    "It could be 'yes' if you had a healthy breakfast.",
+    'Oh, the magic ball just flew out of the window.',
+    "Yes, if we are very precise, but we don't have to be.",
+    'Ask the wizard looking guy next door.',
+    'If I say "yes" do you believe me?',
+     'If I say "no", do you believe me?',
+    'Do I look like I know? Ok, I know. Hehe.',
+    'Let me see... I will tell you later.',
+    'Are you sure you want to ask that?',
+    'Hidden in the mists of magic: It is a mystery.',
+    'I am baking right now, ask later.',
+    'Hmm, I am not in the mood to answer.'
+
+    ]
+
+};
+
+export default answers;
