+++ index.html
@@ -8,11 +8,12 @@
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <meta google-site-verification: google0f303ee427122e7c.html> <!-- google search console -->
         <script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-2676869124180471"
-        crossorigin="anonymous"></script> <!-- google adsense --> 
+        crossorigin="anonymous"></script> <!-- google adsense -->
         <meta name="keywords" content="oracle, oracle game, ask a question, magic 8 ball, random answers, fortune teller,
         mystic responses, question and answer game, fun divination">
         <link href="./resurssit/css/style.css" rel="stylesheet">
-        <script src="main.js" defer></script>
+        <script type="module" src="main.mjs" defer></script>
+ 
     </head>
     <body>
 
@@ -33,11 +34,12 @@
        </div>
        <div class="secondContainer">
         <input class="inputbox" type="text" id="question" name="question" placeholder="ask a question" onkeyup="handleKeyPress(event)">
-         <button class="askButton" type="submit" id="submit" onclick="getValue()">Ask</button>
+        <button class="askButton" type="submit" id="submit" >Ask</button><!--onclick="getValue()  had to delete this when using modules-->
        <div class="userQuestionContainer">
         <p  id="userQuestion"></p> 
        </div>
-       </div>
+        </div>
+ 
 
     </section>   
 
@@ -46,20 +48,27 @@
 <!--errormessage -->
     <div>
       <p class="error">Hey, ask a question!</p>
-      <p class="herbertBubble">Meow meow!</p>
     </div>
 
 
 <!-- cat picture -->
     <div>
         <img  id="herbert" src="./resurssit/kuvat/cat_animation800.gif" height="400" width="600">
+        <p class="herbertBubble">Meow meow!</p>
     </div>
+
+
 
 <!-- oracle picture and answer -->
 
     <section class="oraclecontainer">
      <div >
-    <img class="oracle" src="./resurssit/kuvat/uusimummo.png">
+    <img class="oracleBottom" src="./resurssit/kuvat/uusimummo.png">
+    <img class="oracle" src="./resurssit/kuvat/oracle_no_eyes.png">
+        <div class="eyes">
+            <img class="leftEye" src="./resurssit/kuvat/silma2.png">
+            <img class="rightEye" src="./resurssit/kuvat/silma2.png">
+        </div>
      </div>
 
     <div class="bubble">
@@ -70,7 +79,7 @@
 <!-- Dragon and flames-->
     <section class="dragoncontainer">
         <div>
-             <img class="dragon" src="./resurssit/kuvat/dragon1_transparent.png">
+            <img class="dragon" src="./resurssit/kuvat/dragon1_transparent.png">
         </div>
         <div class="flamescontainer">
             <img class="flame" src="./resurssit/kuvat/flame1_1000.gif">
@@ -78,25 +87,25 @@
             <img class="flame" src="./resurssit/kuvat/flame1_1000.gif">
             <img class="flame" src="./resurssit/kuvat/flame1_1000.gif">
             <img class="flame" src="./resurssit/kuvat/flame1_1000.gif">
-            <img class="flame" src="./resurssit/kuvat/flame1_1000.gif"> 
+            <img class="flame" src="./resurssit/kuvat/flame1_1000.gif">
         </div>
         <div class="yesnocontainer">
-            <button class="yesButton" type="button" onclick="dragonTrigger()">Yes</button>
+            <button class="yesButton" type="button" >Yes</button><!--onclick="dragonTrigger()"-->
             <button class="noButton" type="button"  >No</button>
         </div>
     </section>
+
 
     <footer>
      <div class="counter">   
         <script src="https://static.elfsight.com/platform/platform.js" data-use-service-core defer></script>
         <div class="elfsight-app-d4027d34-28f2-4e27-aed0-2afa894e6f7e" data-elfsight-app-lazy></div>
+        <div class="updated">Updated 4/4/2024</div>
     </div>
-    <div>
-      
-    </div>
+
     </footer>
 
     </body>
 </html>
 
-<!-- C:\Users\Jenny\.vscode\VS_projekti_Omat\JavaScript\The Oracle\index.html -->
+<!-- C:\Users\Jenny\.vscode\VS_projekti_Omat\JavaScript\testialue\index.html  -->
